Using caching library instead of manual caches
