+++ tests/Directory.Build.Targets
@@ -6,7 +6,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="AutofacContrib.NSubstitute" Version="6.1.0" />
+    <PackageReference Include="AutofacContrib.NSubstitute" Version="6.2.0" />
     <PackageReference Include="AutoFixture" Version="4.14.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.7.1" />
     <PackageReference Include="xunit" Version="2.4.1" />
