+++ tests/WWT.Caching.Tests/WWT.Caching.Tests.csproj
@@ -1,0 +1,11 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net48</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\WWT.Caching\WWT.Caching.csproj" />
+  </ItemGroup>
+
+</Project>
