--- tests/WWT.PlateFiles.Caching.Tests/WWT.PlateFiles.Caching.Tests.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net48</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\WWT.PlateFiles.Caching\WWT.PlateFiles.Caching.csproj" />
-  </ItemGroup>
-
-</Project>
