+++ wwt-website.sln
@@ -48,11 +48,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PlateManager", "tools\Plate
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WWT.PlateFiles.Tests", "tests\WWT.PlateFiles.Tests\WWT.PlateFiles.Tests.csproj", "{D34D5F02-3A5B-446C-B02A-68DA5B9BC2AF}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WWT.PlateFiles.Caching", "src\WWT.PlateFiles.Caching\WWT.PlateFiles.Caching.csproj", "{568DE47A-DE03-4B82-836A-C42C4C5A9F84}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WWT.PlateFiles.Caching.Tests", "tests\WWT.PlateFiles.Caching.Tests\WWT.PlateFiles.Caching.Tests.csproj", "{A34C0E73-F384-4FBA-AE89-3ECCC48892C1}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WWT.Imaging", "src\WWT.Imaging\WWT.Imaging.csproj", "{33640247-520C-4F61-9055-B02CD8B66B8A}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WWT.Caching", "src\WWT.Caching\WWT.Caching.csproj", "{CEBAB32C-42A7-43F8-AD7B-2521ED5AB12E}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WWT.Caching.Tests", "tests\WWT.Caching.Tests\WWT.Caching.Tests.csproj", "{3FBE3897-A98D-4066-8B6C-3F603A23DFB2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -100,18 +100,18 @@ Global
 		{D34D5F02-3A5B-446C-B02A-68DA5B9BC2AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D34D5F02-3A5B-446C-B02A-68DA5B9BC2AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D34D5F02-3A5B-446C-B02A-68DA5B9BC2AF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{568DE47A-DE03-4B82-836A-C42C4C5A9F84}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{568DE47A-DE03-4B82-836A-C42C4C5A9F84}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{568DE47A-DE03-4B82-836A-C42C4C5A9F84}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{568DE47A-DE03-4B82-836A-C42C4C5A9F84}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A34C0E73-F384-4FBA-AE89-3ECCC48892C1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A34C0E73-F384-4FBA-AE89-3ECCC48892C1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A34C0E73-F384-4FBA-AE89-3ECCC48892C1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A34C0E73-F384-4FBA-AE89-3ECCC48892C1}.Release|Any CPU.Build.0 = Release|Any CPU
 		{33640247-520C-4F61-9055-B02CD8B66B8A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{33640247-520C-4F61-9055-B02CD8B66B8A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{33640247-520C-4F61-9055-B02CD8B66B8A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{33640247-520C-4F61-9055-B02CD8B66B8A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CEBAB32C-42A7-43F8-AD7B-2521ED5AB12E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CEBAB32C-42A7-43F8-AD7B-2521ED5AB12E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CEBAB32C-42A7-43F8-AD7B-2521ED5AB12E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CEBAB32C-42A7-43F8-AD7B-2521ED5AB12E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3FBE3897-A98D-4066-8B6C-3F603A23DFB2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3FBE3897-A98D-4066-8B6C-3F603A23DFB2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3FBE3897-A98D-4066-8B6C-3F603A23DFB2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3FBE3897-A98D-4066-8B6C-3F603A23DFB2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -127,9 +127,9 @@ Global
 		{C5E54B43-A19F-4021-875C-E8769D89EB1A} = {EFF21404-6CEA-49EB-9402-C6643AC2564F}
 		{94CB88BE-875F-451D-8BB2-36327DA90708} = {45CC6DAC-BC15-4EEC-ABD3-34EC42E93221}
 		{D34D5F02-3A5B-446C-B02A-68DA5B9BC2AF} = {EFF21404-6CEA-49EB-9402-C6643AC2564F}
-		{568DE47A-DE03-4B82-836A-C42C4C5A9F84} = {F12286DB-AF41-4661-A1C6-EB3259CB9B41}
-		{A34C0E73-F384-4FBA-AE89-3ECCC48892C1} = {EFF21404-6CEA-49EB-9402-C6643AC2564F}
 		{33640247-520C-4F61-9055-B02CD8B66B8A} = {F12286DB-AF41-4661-A1C6-EB3259CB9B41}
+		{CEBAB32C-42A7-43F8-AD7B-2521ED5AB12E} = {F12286DB-AF41-4661-A1C6-EB3259CB9B41}
+		{3FBE3897-A98D-4066-8B6C-3F603A23DFB2} = {EFF21404-6CEA-49EB-9402-C6643AC2564F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {80DABDD3-0D48-43AD-893B-B09A38FCFC72}
